add basic amortization table

+++ Domain/Contracts/ILoanService.cs
@@ -8,6 +8,6 @@ namespace Domain.Contracts
 {
     public interface ILoanService
     {
-        Loan GetAmortization(decimal capital, decimal bankRate, int quantityAliquot, int modality, DateTimeOffset startDate);
+        Loan GetAmortization(decimal capital, double bankRate, int quantityAliquot, int modality, DateTimeOffset startDate);
     }
 }
